Split sanity tests into tests_basic.py
