+++ test/cmd_line/c_graph.dot
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<graphml xmlns="http://graphml.graphdrawing.org/xmlns" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://graphml.graphdrawing.org/xmlns http://graphml.graphdrawing.org/xmlns/1.0/graphml.xsd">
+  <key id="key0" for="edge" attr.name="line" attr.type="int" />
+  <key id="key1" for="node" attr.name="name" attr.type="string" />
+  <graph id="G" edgedefault="directed" parse.nodeids="free" parse.edgeids="canonical" parse.order="nodesfirst">
+    <node id="n0">
+      <data key="key1">inc/lib2/f_4.h</data>
+    </node>
+    <node id="n1">
+      <data key="key1">inc/lib2/f_1.h</data>
+    </node>
+    <node id="n2">
+      <data key="key1">inc/lib/f_3.h</data>
+    </node>
+    <node id="n3">
+      <data key="key1">inc/lib/f_1.h</data>
+    </node>
+    <node id="n4">
+      <data key="key1">inc/lib/f_2.h</data>
+    </node>
+    <node id="n5">
+      <data key="key1">src/f_2.c</data>
+    </node>
+    <node id="n6">
+      <data key="key1">src/f_1.c</data>
+    </node>
+    <node id="n7">
+      <data key="key1">src/f_3.c</data>
+    </node>
+    <node id="n8">
+      <data key="key1">iostream</data>
+    </node>
+    <node id="n9">
+      <data key="key1">../non/existing.h</data>
+    </node>
+    <edge id="e0" source="n2" target="n3">
+      <data key="key0">1</data>
+    </edge>
+    <edge id="e1" source="n3" target="n2">
+      <data key="key0">4</data>
+    </edge>
+    <edge id="e2" source="n3" target="n0">
+      <data key="key0">8</data>
+    </edge>
+    <edge id="e3" source="n4" target="n3">
+      <data key="key0">1</data>
+    </edge>
+    <edge id="e4" source="n5" target="n1">
+      <data key="key0">1</data>
+    </edge>
+    <edge id="e5" source="n6" target="n3">
+      <data key="key0">1</data>
+    </edge>
+    <edge id="e6" source="n6" target="n4">
+      <data key="key0">2</data>
+    </edge>
+    <edge id="e7" source="n6" target="n8">
+      <data key="key0">5</data>
+    </edge>
+    <edge id="e8" source="n7" target="n2">
+      <data key="key0">1</data>
+    </edge>
+    <edge id="e9" source="n7" target="n8">
+      <data key="key0">4</data>
+    </edge>
+    <edge id="e10" source="n7" target="n9">
+      <data key="key0">7</data>
+    </edge>
+  </graph>
+</graphml>
