+++ test/cmd_line/ruby_graph.dot
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<graphml xmlns="http://graphml.graphdrawing.org/xmlns" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://graphml.graphdrawing.org/xmlns http://graphml.graphdrawing.org/xmlns/1.0/graphml.xsd">
+  <key id="key0" for="edge" attr.name="line" attr.type="int" />
+  <key id="key1" for="node" attr.name="name" attr.type="string" />
+  <graph id="G" edgedefault="directed" parse.nodeids="free" parse.edgeids="canonical" parse.order="nodesfirst">
+    <node id="n0">
+      <data key="key1">motorcycle_test.rb</data>
+    </node>
+    <node id="n1">
+      <data key="key1">lib/motorcycle.rb</data>
+    </node>
+    <edge id="e0" source="n0" target="n1">
+      <data key="key0">1</data>
+    </edge>
+  </graph>
+</graphml>
