+++ test/cmd_line/test.py
@@ -23,14 +23,16 @@ import pygraphviz as pgv
 import pygraphml as pgml
 
 from subprocess import Popen, PIPE
-from os.path import abspath
+from os.path import abspath, join
 
 
 class GardenerTestCases(unittest.TestCase):
 
     # default paths
-    G_PATH = "../../build/include_gardener"
+    M_PATH = "."
     T_PATH = "../test_files/"
+    G_PATH = "../../build/include_gardener"
+    LANGUAGE = "c"
 
     def setUp(self):
         """Nothing to setup here"""
@@ -44,42 +46,8 @@ class GardenerTestCases(unittest.TestCase):
 
             A PyGraphml graph is returned.
         """
-        g = pgml.Graph()
-
-        f_1_c = g.add_node("1")
-        f_2_c = g.add_node("2")
-        f_3_c = g.add_node("3")
-        existing_h = g.add_node("4")
-        iostream = g.add_node("5")
-        lib_f_1_h = g.add_node("6")
-        lib_f_2_h = g.add_node("7")
-        lib_f_3_h = g.add_node("8")
-        lib2_f_4_h = g.add_node("9")
-        lib2_f_1_h = g.add_node("10")
-
-        f_1_c['key1'] = 'src/f_1.c'
-        f_2_c['key1'] = 'src/f_2.c'
-        f_3_c['key1'] = 'src/f_3.c'
-        existing_h['key1'] = '../non/existing.h'
-        iostream['key1'] = 'iostream'
-        lib_f_1_h['key1'] = 'inc/lib/f_1.h'
-        lib_f_2_h['key1'] = 'inc/lib/f_2.h'
-        lib_f_3_h['key1'] = 'inc/lib/f_3.h'
-        lib2_f_4_h['key1'] = 'inc/lib2/f_4.h'
-        lib2_f_1_h['key1'] = 'inc/lib2/f_1.h'
-
-        g.add_edge(f_3_c, existing_h)
-        g.add_edge(f_3_c, iostream)
-        g.add_edge(f_2_c, lib2_f_1_h)
-        g.add_edge(f_1_c, iostream)
-        g.add_edge(f_1_c, lib_f_2_h)
-        g.add_edge(f_1_c, lib_f_1_h)
-        g.add_edge(f_3_c, lib_f_3_h)
-        g.add_edge(lib_f_2_h, lib_f_1_h)
-        g.add_edge(lib_f_1_h, lib_f_3_h)
-        g.add_edge(lib_f_1_h, lib2_f_4_h)
-        g.add_edge(lib_f_3_h, lib_f_1_h)
-        return g
+        parser = pgml.GraphMLParser()
+        return parser.parse(os.path.join(self.M_PATH, self.LANGUAGE + '_graph.dot'))
 
     def compare(self, G1, G2):
         """ Compares two PyGraphml graphs by using PyUnittest's
@@ -153,80 +121,13 @@ class GardenerTestCases(unittest.TestCase):
         # get the result from the system call:
         return parser.parse(temp.name)
 
-    def test_showsUnrecognisedOption(self):
-        """Tests "include_gardener --xyz"
-
-        The test expects an unrecognised option information.
-        """
-        pipe = Popen([self.G_PATH, "--xyz"], stderr=PIPE)
-        result = pipe.communicate()[1].decode("utf-8")
-        self.assertEqual(result, "unrecognised option '--xyz'\n")
-
-    def test_printHelp1(self):
-        """Tests "include_gardener -h"
-
-        The test expects at least 'Options:' in stdout.
-        """
-        pipe = Popen([self.G_PATH, "-h"], stdout=PIPE)
-        result = pipe.communicate()[0].decode("utf-8")
-        self.assertIn("Options:", result)
-
-    def test_printHelp2(self):
-        """Tests "include_gardener --help"
-
-        The test expects at least 'Options:' in stdout.
-        """
-        pipe = Popen([self.G_PATH, "--help"], stdout=PIPE)
-        result = pipe.communicate()[0].decode("utf-8")
-        self.assertIn("Options:", result)
-
-    def test_Version1(self):
-        """Tests "include_gardener -v"
-
-        The test expects at least 'Version' in stdout.
-        """
-        pipe = Popen([self.G_PATH, "-v"], stdout=PIPE)
-        result = pipe.communicate()[0].decode("utf-8")
-        self.assertIn("Version", result)
-
-    def test_Version2(self):
-        """Tests "include_gardener --version"
-
-        The test expects at least 'Version' in stdout.
-        """
-        pipe = Popen([self.G_PATH, "--version"], stdout=PIPE)
-        result = pipe.communicate()[0].decode("utf-8")
-        self.assertIn("Version", result)
-
-    def test_UnsuportedLanguage(self):
-        """ Tests if an error message is given if an unsupported language
-            is selected
-        """
-        pipe = Popen([self.G_PATH, "-l", "x-lang", self.T_PATH], stderr=PIPE)
-        result = pipe.communicate()[1].decode("utf-8")
-        self.assertIn("Error", result)
-        self.assertIn("x-lang", result)
-
-    def test_SimpleCallWithSinglePath(self):
-        """ Tests "include_gardener test_files"
-
-        The test expects that the result can be read by pygraphviz
-        and that there is at least one node.
-        """
-        G = self.dot_gardener_call(['-j', '1'])
-
-        # the first node shall not be None ...
-        n = G.get_node(1)
-        self.assertNotEqual(n, None)
-
-    def test_SimpleCallWithSinglePath_GraphmlOutput(self):
-        """ Tests "include_gardener test_files -f xml -I test_files/inc"
+    def test_CallAndCompareOutput(self):
+        """ Tests "include_gardener -f xml -l <language> -I <include_files> <test_files>"
 
         The test expects that the result can be read by graphml
         and that there is at least one node.
         """
-        print(self.T_PATH + 'inc/')
-        g1 = self.graphml_gardener_call(['-I', self.T_PATH + '/inc/'])
+        g1 = self.graphml_gardener_call(['-l', self.LANGUAGE, '-I', self.I_PATH])
 
         # get a reference graph
         g2 = self.build_reference_graph()
@@ -234,93 +135,16 @@ class GardenerTestCases(unittest.TestCase):
         # both graphs shall be the same
         self.compare(g1, g2)
 
-    def test_LevelOption(self):
-        """ Tests "include_gardener test_files -f xml" once without -L,
-            once with -L 0, once with -L 1 and once with -L 2.
-
-        The test expects that -L 0 counts no nodes, -L 1 counts just the
-        nodes of the result of processing the files in src, and L 2 the
-        same than wihtout the -L option.
-        """
-        g1 = self.graphml_gardener_call([])
-        g2 = self.graphml_gardener_call(['-L', '0'])
-        g3 = self.graphml_gardener_call(['-L', '1'])
-        g4 = self.graphml_gardener_call(['-L', '2'])
-
-        self.assertEqual(len(g2.nodes()), 0)
-        self.assertEqual(len(g3.nodes()), 9)
-        self.assertEqual(len(g4.nodes()), 15)
-        self.compare(g1, g4)
-
-    def test_ThreadOption(self):
-        """ Tests "include_gardener test_files -f xml" with -j ranging
-            from 1 to 5, each 10 times. Each call should produce the same
-            result.
-        """
-
-        g1 = None
-        for j in range(1, 5):
-            for cnt in range(0, 10):
-                if not g1:
-                    # initialize reference
-                    g1 = self.graphml_gardener_call(['-I', self.T_PATH +
-                                                     '/inc/', '-j', str(j)])
-                else:
-                    g2 = self.graphml_gardener_call(['-I', self.T_PATH +
-                                                     '/inc/', '-j', str(j)])
-                    self.compare(g1, g2)
-
-    def test_ThreadNullOption(self):
-        """ Tests "include_gardener test_files -f xml -j 0" which should fail.
-        """
-        pipe = Popen([self.G_PATH, self.T_PATH,
-                      '-f', 'xml', '-j', '0'], stderr=PIPE)
-        graphml_str = pipe.communicate()[1].decode("utf-8")
-        self.assertNotEqual(len(graphml_str), 0)
-        self.assertIn("Error", graphml_str)
-
-    def test_ExcludeOption(self):
-        """ Tests "include_gardener test_files -f xml"
-        once without the -e option, once with -e _1\.h and
-        once with -e _1\.h -e _1\.c.
-        The test expects that the first includes all files,
-        the second call one file less and the third two files less.
-        """
-        g1 = self.graphml_gardener_call(['-I', self.T_PATH + '/inc/'])
-        g2 = self.graphml_gardener_call(['-I', self.T_PATH + '/inc/',
-                                         '-e', '_2\.c'])
-        g3 = self.graphml_gardener_call(['-I', self.T_PATH + '/inc/',
-                                         '-e', '_2\.c', '-e', '_3\.c'])
-
-        self.assertEqual(len(g1.nodes()), 10)
-        self.assertEqual(len(g2.nodes()), 9)
-
-        # note: there are two three lodes less than in g1:
-        # - f_2.c
-        # - f_3.c
-        # - non/existing.h (which is referenced in f_3.c)
-        self.assertEqual(len(g3.nodes()), 7)
-
-    def test_OutputFile(self):
-        """ Tests if the gardener can write into an output file.
-
-        The test expects that there is at least one node.
-        """
-        temp = tempfile.NamedTemporaryFile()
-        self.graphml_gardener_call(['-o', temp.name])
-        parser = pgml.GraphMLParser()
-
-        # get the result from the system call:
-        g = parser.parse(temp.name)
-        temp = None
-        self.assertNotEqual(len(g.nodes()), 0)
-
 
 if __name__ == "__main__":
     print(sys.argv)
-    if len(sys.argv) > 2:
+    if len(sys.argv) > 4:
         GardenerTestCases.T_PATH = abspath(sys.argv.pop())
+        GardenerTestCases.I_PATH = abspath(sys.argv.pop())
         GardenerTestCases.G_PATH = abspath(sys.argv.pop())
+        GardenerTestCases.M_PATH = abspath(sys.argv.pop())
+        GardenerTestCases.LANGUAGE = sys.argv.pop()
+
     import xmlrunner
     out_dir = abspath(sys.argv.pop())
     unittest.main(testRunner=xmlrunner.XMLTestRunner(output=out_dir))
